Deprecate 0x111
Note

Deprecated: This packet is no longer used.

This packet was sent by the client when requesting to lock their style. This packet is no longer used as the style lock system has undergone several changes since this packet was first added. Other packets are now used to handle this feature.

Co-Authored-By: atom0s <atom0s@users.noreply.github.com>